backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
asyncpg
pydantic-settings
python-jose[cryptography]
passlib[bcrypt]
httpx
stripe
google-auth
google-auth-oauthlib
google-api-python-client
playwright
